refactor(results): extract shared TopBarriersContent component
Eliminates the live vs CRP-2026 drift class on /results/top-barriers
and /results/crp-2026/top-barriers by introducing a single
<TopBarriersContent variant="live" | "crp" data={...} /> component, the
same pattern already used by FactorAnalysisContent and ValidationPageContent.

Both page.tsx files become 16-line shells that import their tier's JSON
and the shared component. The two pages had drifted to 174 diff lines
including:

- accessibility regression on the CRP page (missing scope="col"/scope="row"
  attributes on table headers and the leftmost row label cell)
- divergent guard logic on the "omitted Mean Rank" explanatory sentence
  (live had dataAvailable && hasOmittedMeanRank, CRP had only the latter)
- stale prettier formatting on the CRP page

The refactor uses the live page as the authoritative source for
component logic and a11y attributes, so the CRP page automatically
inherits all those improvements. Variant-specific bits (title, intro
prose, cross-reference links, badges, JSON file path label, source-data
prose) are gated by the variant prop.

No behavioral change for end users beyond the CRP page now having the
correct table-cell scoping and the same guard behavior as live.

Tests: 606/606 pass. Build clean.
